--- a/src/@types/cozy-client.d.ts
+++ b/src/@types/cozy-client.d.ts
@@ -176,10 +176,14 @@ declare module 'cozy-client' {
       queryDefinition: QueryDefinition,
       options?: QueryOptions
     ) => Promise<QueryResult>
-    queryAll: <T>(queryDefinition: QueryDefinition) => Promise<T>
+    queryAll: <T>(
+      queryDefinition: QueryDefinition,
+      options?: QueryOptions
+    ) => Promise<T>
     links: CozyLink[]
     capabilities: ClientCapabilities
     registerPlugin: (Plugin: Function, options: unknown) => void
+    getCollectionFromState: (doctype: string) => unknown
   }
 
   export const createMockClient = (options?: ClientOptions): CozyClient =>