--- a/src/search/consts.ts
+++ b/src/search/consts.ts
@@ -29,6 +29,7 @@ export const CONTACTS_DOCTYPE = 'io.cozy.contacts'
 export const APPS_DOCTYPE = 'io.cozy.apps'
 
 export const TYPE_DIRECTORY = 'directory'
+export const TYPE_FILE = 'file'
 export const ROOT_DIR_ID = 'io.cozy.files.root-dir'
 export const TRASH_DIR_ID = 'io.cozy.files.trash-dir'
 export const SHARED_DRIVES_DIR_ID = 'io.cozy.files.shared-drives-dir'