Since https://github.com/cozy/cozy-web-data-proxy/pull/6, we noticed a severe performance degradation when performing search.

We measured the following:
- Flexsearch search: <ms
- Query cozy-client store to get file paths: ~120ms for 60 files.
- Results normalization: ~3ms

We used to make multiple `getById` queries to get file path, each of them taking few ms. So we tried to use a `getByIds` instead, and got ~70ms per query, which is better, but still slow.
Then, we tried the following:
```
  const allFiles = client.getCollectionFromState(FILES_DOCTYPE) as IOCozyFile[]
  const dirs = allFiles.filter(file => file.type === TYPE_DIRECTORY)
  return dirs.filter(dir => dirIds.includes(dir._id))
```

This takes 2ms.

We should eventually investigate why the store queries are so slow, but for now let's use this trick.

:warning: This is unrelated to this PR, but we also noticed a dramatic performance drop when using Firefox: the postmessage communication takes ~400ms for 100 search results vs 02.ms with Chrome